requirements.txt
numpy==1.21.2  # https://pypi.org/project/numpy/
websocket_client==0.57.0 # https://pypi.org/project/websocket_client/
python-binance==0.7.5 # https://pypi.org/project/python-binance/
config==0.5.0.post0 # https://pypi.org/project/config/
pytest==6.2.1 # https://pypi.org/project/pytest/
pytest-cov==2.10.1 # https://pypi.org/project/pytest-cov/
sqlalchemy==1.3.23 #
mpl_finance==0.10.1 #
pandas==1.2.3 #
plotly

pytest
black==19.10b0
python-dotenv